Migrate build scripts to Nuke build project.

+++ sources/FFmpegSharp.Interop/avdevice/AVAppToDevMessageType.cs
@@ -21,17 +21,17 @@ namespace FFmpegSharp.Interop
     [NativeTypeName("int")]
     public enum AVAppToDevMessageType : uint
     {
-        AV_APP_TO_DEV_NONE = (unchecked(('E') | (('N') << 8) | (('O') << 16) | ((uint)('N') << 24))),
-        AV_APP_TO_DEV_WINDOW_SIZE = (unchecked(('M') | (('O') << 8) | (('E') << 16) | ((uint)('G') << 24))),
-        AV_APP_TO_DEV_WINDOW_REPAINT = (unchecked(('A') | (('P') << 8) | (('E') << 16) | ((uint)('R') << 24))),
-        AV_APP_TO_DEV_PAUSE = (unchecked((' ') | (('U') << 8) | (('A') << 16) | ((uint)('P') << 24))),
-        AV_APP_TO_DEV_PLAY = (unchecked(('Y') | (('A') << 8) | (('L') << 16) | ((uint)('P') << 24))),
-        AV_APP_TO_DEV_TOGGLE_PAUSE = (unchecked(('T') | (('U') << 8) | (('A') << 16) | ((uint)('P') << 24))),
-        AV_APP_TO_DEV_SET_VOLUME = (unchecked(('L') | (('O') << 8) | (('V') << 16) | ((uint)('S') << 24))),
-        AV_APP_TO_DEV_MUTE = (unchecked(('T') | (('U') << 8) | (('M') << 16) | ((uint)(' ') << 24))),
-        AV_APP_TO_DEV_UNMUTE = (unchecked(('T') | (('U') << 8) | (('M') << 16) | ((uint)('U') << 24))),
-        AV_APP_TO_DEV_TOGGLE_MUTE = (unchecked(('T') | (('U') << 8) | (('M') << 16) | ((uint)('T') << 24))),
-        AV_APP_TO_DEV_GET_VOLUME = (unchecked(('L') | (('O') << 8) | (('V') << 16) | ((uint)('G') << 24))),
-        AV_APP_TO_DEV_GET_MUTE = (unchecked(('T') | (('U') << 8) | (('M') << 16) | ((uint)('G') << 24))),
+        AV_APP_TO_DEV_NONE = (unchecked(((sbyte)('E')) | (((sbyte)('N')) << 8) | (((sbyte)('O')) << 16) | ((uint)('N') << 24))),
+        AV_APP_TO_DEV_WINDOW_SIZE = (unchecked(((sbyte)('M')) | (((sbyte)('O')) << 8) | (((sbyte)('E')) << 16) | ((uint)('G') << 24))),
+        AV_APP_TO_DEV_WINDOW_REPAINT = (unchecked(((sbyte)('A')) | (((sbyte)('P')) << 8) | (((sbyte)('E')) << 16) | ((uint)('R') << 24))),
+        AV_APP_TO_DEV_PAUSE = (unchecked(((sbyte)(' ')) | (((sbyte)('U')) << 8) | (((sbyte)('A')) << 16) | ((uint)('P') << 24))),
+        AV_APP_TO_DEV_PLAY = (unchecked(((sbyte)('Y')) | (((sbyte)('A')) << 8) | (((sbyte)('L')) << 16) | ((uint)('P') << 24))),
+        AV_APP_TO_DEV_TOGGLE_PAUSE = (unchecked(((sbyte)('T')) | (((sbyte)('U')) << 8) | (((sbyte)('A')) << 16) | ((uint)('P') << 24))),
+        AV_APP_TO_DEV_SET_VOLUME = (unchecked(((sbyte)('L')) | (((sbyte)('O')) << 8) | (((sbyte)('V')) << 16) | ((uint)('S') << 24))),
+        AV_APP_TO_DEV_MUTE = (unchecked(((sbyte)('T')) | (((sbyte)('U')) << 8) | (((sbyte)('M')) << 16) | ((uint)(' ') << 24))),
+        AV_APP_TO_DEV_UNMUTE = (unchecked(((sbyte)('T')) | (((sbyte)('U')) << 8) | (((sbyte)('M')) << 16) | ((uint)('U') << 24))),
+        AV_APP_TO_DEV_TOGGLE_MUTE = (unchecked(((sbyte)('T')) | (((sbyte)('U')) << 8) | (((sbyte)('M')) << 16) | ((uint)('T') << 24))),
+        AV_APP_TO_DEV_GET_VOLUME = (unchecked(((sbyte)('L')) | (((sbyte)('O')) << 8) | (((sbyte)('V')) << 16) | ((uint)('G') << 24))),
+        AV_APP_TO_DEV_GET_MUTE = (unchecked(((sbyte)('T')) | (((sbyte)('U')) << 8) | (((sbyte)('M')) << 16) | ((uint)('G') << 24))),
     }
 }
